Move pkg to internal
Those files are not meant to be used outside of lemonldap-ng-controller
