+++ internal/signals/signal.go
